app/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
geopy
plotly
python-dotenv
requests
pydantic
Pillow
